Extract AssistantHandle
